Fix JwkSerializer and tests
