+++ sdk/src/sharedTest/java/com/microsoft/did/sdk/backup/content/microsoft2020/TestVcMetaData.kt
@@ -8,4 +8,4 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 @SerialName("TestVcMetaData")
-class TestVcMetaData(override val displayContract: DisplayContract) : VcMetadata()
+data class TestVcMetaData(override val displayContract: DisplayContract) : VcMetadata()
