+++ sdk/src/test/java/com/microsoft/did/sdk/BackupServiceTest.kt
@@ -2,7 +2,7 @@
 
 package com.microsoft.did.sdk
 
-import android.util.VerifiableCredentialUtil
+import android.util.BackupTestUtil
 import com.microsoft.did.sdk.backup.BackupParser
 import com.microsoft.did.sdk.backup.container.jwe.JwePasswordProtectedBackupData
 import com.microsoft.did.sdk.backup.container.jwe.JwePasswordProtectionMethod
@@ -11,20 +11,16 @@ import com.microsoft.did.sdk.backup.content.microsoft2020.Microsoft2020BackupPro
 import com.microsoft.did.sdk.backup.content.microsoft2020.Microsoft2020UnprotectedBackup
 import com.microsoft.did.sdk.backup.content.microsoft2020.RawIdentifierConverter
 import com.microsoft.did.sdk.backup.content.microsoft2020.TestVcMetaData
-import com.microsoft.did.sdk.backup.content.microsoft2020.VcMetadata
 import com.microsoft.did.sdk.backup.content.microsoft2020.WalletMetadata
-import com.microsoft.did.sdk.credential.service.models.contracts.display.DisplayContract
 import com.microsoft.did.sdk.util.controlflow.Result
 import com.microsoft.did.sdk.util.defaultTestSerializer
 import kotlinx.coroutines.runBlocking
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.Test
 
 class BackupServiceTest {
-    private val identifierRepository = VerifiableCredentialUtil.getMockIdentifierRepository()
-    private val keyStore = VerifiableCredentialUtil.getMockKeyStore()
+    private val identifierRepository = BackupTestUtil.getMockIdentifierRepository()
+    private val keyStore = BackupTestUtil.getMockKeyStore()
 
     private val jweBackupFactory = BackupParser()
     private val microsoft2020BackupProcessor = Microsoft2020BackupProcessor(
@@ -38,11 +34,11 @@ class BackupServiceTest {
     private val service = BackupService(jweBackupFactory, backupProcessorFactory, defaultTestSerializer)
     private val password = "Big complex passsword you'll never be able to guess"
 
-    private val vcMetadata = TestVcMetaData(VerifiableCredentialUtil.testDisplayContract)
+    private val vcMetadata = TestVcMetaData(BackupTestUtil.testDisplayContract)
 
     private val backup = Microsoft2020UnprotectedBackup(
         WalletMetadata(),
-        listOf(Pair(VerifiableCredentialUtil.testVerifiedCredential, vcMetadata))
+        listOf(Pair(BackupTestUtil.testVerifiedCredential, vcMetadata))
     )
 
     @Test
@@ -52,7 +48,7 @@ class BackupServiceTest {
             val protectedBackupData = (service.exportBackup(backup, protectionMethod) as Result.Success).payload
             assertThat(protectedBackupData).isInstanceOf(JwePasswordProtectedBackupData::class.java)
 
-            val actualBackup = service.importBackup(protectedBackupData, protectionMethod)
+            val actualBackup = (service.importBackup(protectedBackupData, protectionMethod) as Result.Success).payload
             assertThat(actualBackup).isEqualToComparingFieldByFieldRecursively(backup)
         }
     }
