+++ sdk/src/test/java/com/microsoft/did/sdk/backup/content/microsoft2020/RawIdentifierConverterTest.kt
@@ -2,7 +2,7 @@
 
 package com.microsoft.did.sdk.backup.content.microsoft2020
 
-import android.util.VerifiableCredentialUtil
+import android.util.BackupTestUtil
 import com.nimbusds.jose.jwk.KeyOperation
 import com.nimbusds.jose.jwk.KeyUse
 import kotlinx.coroutines.runBlocking
@@ -13,19 +13,19 @@ import kotlin.test.assertTrue
 import kotlin.test.fail
 
 class RawIdentifierConverterTest {
-    private val identifierRepository = VerifiableCredentialUtil.getMockIdentifierRepository()
-    private val keyStore = VerifiableCredentialUtil.getMockKeyStore()
+    private val identifierRepository = BackupTestUtil.getMockIdentifierRepository()
+    private val keyStore = BackupTestUtil.getMockKeyStore()
     private val rawIdentifierUtility = RawIdentifierConverter(identifierRepository, keyStore)
 
     @Test
     fun parseRawIdentifierTest() {
-        val actual = rawIdentifierUtility.parseRawIdentifier(VerifiableCredentialUtil.rawIdentifier)
-        assertEquals(VerifiableCredentialUtil.testIdentifer, actual.first)
+        val actual = rawIdentifierUtility.parseRawIdentifier(BackupTestUtil.rawIdentifier)
+        assertEquals(BackupTestUtil.testIdentifer, actual.first)
         assertEquals(4, actual.second.size, "expected four distinct keys")
-        assertTrue(actual.second.contains(VerifiableCredentialUtil.encryptKey))
-        assertTrue(actual.second.contains(VerifiableCredentialUtil.signKey))
-        assertTrue(actual.second.contains(VerifiableCredentialUtil.updateKey))
-        assertTrue(actual.second.contains(VerifiableCredentialUtil.recoverKey))
+        assertTrue(actual.second.contains(BackupTestUtil.encryptKey))
+        assertTrue(actual.second.contains(BackupTestUtil.signKey))
+        assertTrue(actual.second.contains(BackupTestUtil.updateKey))
+        assertTrue(actual.second.contains(BackupTestUtil.recoverKey))
     }
 
     @Test
@@ -33,13 +33,13 @@ class RawIdentifierConverterTest {
         runBlocking {
             val actual = rawIdentifierUtility.getAllIdentifiers()
             assertEquals(1, actual.size, "expected one DID")
-            assertEquals(VerifiableCredentialUtil.rawIdentifier.id, actual[0].id)
+            assertEquals(BackupTestUtil.rawIdentifier.id, actual[0].id)
             validateRawIdentifier(actual[0])
         }
     }
 
     private fun validateRawIdentifier(rawIdentity: RawIdentity?) {
-        val expected = VerifiableCredentialUtil.rawIdentifier
+        val expected = BackupTestUtil.rawIdentifier
         assertNotNull(rawIdentity, "Failed to form rawIdentifier")
         assertEquals(expected.id, rawIdentity.id, "DID does not match")
         assertEquals(expected.name, rawIdentity.name, "Name does not match")
@@ -47,16 +47,16 @@ class RawIdentifierConverterTest {
         assertEquals(expected.updateKey, rawIdentity.updateKey, "update key id does not match")
         rawIdentity.keys.forEach {
             when (it.keyID) {
-                VerifiableCredentialUtil.recoverKey.keyID,
-                VerifiableCredentialUtil.updateKey.keyID -> {
+                BackupTestUtil.recoverKey.keyID,
+                BackupTestUtil.updateKey.keyID -> {
                     // do nothing, we've checked above
                 }
-                VerifiableCredentialUtil.signKey.keyID -> {
+                BackupTestUtil.signKey.keyID -> {
                     val correctUse = it.keyUse == KeyUse.SIGNATURE
                     val correctOps = it.keyOperations?.contains(KeyOperation.SIGN)
                     assertTrue(correctOps == true || correctUse, "Did not find correct key_ops or use")
                 }
-                VerifiableCredentialUtil.encryptKey.keyID -> {
+                BackupTestUtil.encryptKey.keyID -> {
                     val correctUse = it.keyUse == KeyUse.ENCRYPTION
                     val correctOps = it.keyOperations?.contains(KeyOperation.UNWRAP_KEY)
                     assertTrue(correctOps == true || correctUse, "Did not find correct key_ops or use")
